add common testing module
