+++ instrumentation/opensearch/opensearch-rest-1.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/opensearch/rest/v1_0/OpenSearchRestTest.java
@@ -5,26 +5,9 @@
 
 package io.opentelemetry.javaagent.instrumentation.opensearch.rest.v1_0;
 
-import static io.opentelemetry.instrumentation.testing.junit.db.SemconvStabilityUtil.maybeStable;
-import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
-import static io.opentelemetry.semconv.HttpAttributes.HTTP_REQUEST_METHOD;
-import static io.opentelemetry.semconv.HttpAttributes.HTTP_RESPONSE_STATUS_CODE;
-import static io.opentelemetry.semconv.NetworkAttributes.NETWORK_PROTOCOL_VERSION;
-import static io.opentelemetry.semconv.ServerAttributes.SERVER_ADDRESS;
-import static io.opentelemetry.semconv.ServerAttributes.SERVER_PORT;
-import static io.opentelemetry.semconv.UrlAttributes.URL_FULL;
-import static io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DB_OPERATION;
-import static io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DB_STATEMENT;
-import static io.opentelemetry.semconv.incubating.DbIncubatingAttributes.DB_SYSTEM;
-
-import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
-import java.io.IOException;
-import java.security.KeyManagementException;
-import java.security.KeyStoreException;
-import java.security.NoSuchAlgorithmException;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicReference;
+import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
+import io.opentelemetry.javaagent.instrumentation.opensearch.rest.AbstractOpenSearchRestTest;
 import javax.net.ssl.SSLContext;
 import org.apache.http.HttpHost;
 import org.apache.http.auth.AuthScope;
@@ -34,38 +17,21 @@ import org.apache.http.conn.ssl.NoopHostnameVerifier;
 import org.apache.http.conn.ssl.TrustAllStrategy;
 import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.ssl.SSLContextBuilder;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
-import org.opensearch.client.Request;
 import org.opensearch.client.Response;
-import org.opensearch.client.ResponseListener;
 import org.opensearch.client.RestClient;
-import org.opensearch.testcontainers.OpensearchContainer;
-import org.testcontainers.utility.DockerImageName;
 
-@SuppressWarnings("deprecation") // using deprecated semconv
-public class OpenSearchRestTest {
+class OpenSearchRestTest extends AbstractOpenSearchRestTest {
   @RegisterExtension
   static final AgentInstrumentationExtension testing = AgentInstrumentationExtension.create();
 
-  static OpensearchContainer opensearch;
-  static RestClient client;
+  @Override
+  protected InstrumentationExtension getTesting() {
+    return testing;
+  }
 
-  static HttpHost httpHost;
-
-  @BeforeAll
-  static void setUpOpenSearch()
-      throws NoSuchAlgorithmException, KeyStoreException, KeyManagementException {
-
-    opensearch =
-        new OpensearchContainer(DockerImageName.parse("opensearchproject/opensearch:1.3.6"))
-            .withSecurityEnabled();
-    // limit memory usage
-    opensearch.withEnv("OPENSEARCH_JAVA_OPTS", "-Xmx256m -Xms256m");
-    opensearch.start();
-
+  @Override
+  protected RestClient buildRestClient() throws Exception {
     CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
     credentialsProvider.setCredentials(
         AuthScope.ANY,
@@ -74,8 +40,8 @@ public class OpenSearchRestTest {
     SSLContext sslContext =
         SSLContextBuilder.create().loadTrustMaterial(null, new TrustAllStrategy()).build();
 
-    httpHost = HttpHost.create(opensearch.getHttpHostAddress());
-    client =
+    HttpHost httpHost = HttpHost.create(opensearch.getHttpHostAddress());
+    return
         RestClient.builder(httpHost)
             .setHttpClientConfigCallback(
                 httpClientBuilder ->
@@ -88,104 +54,8 @@ public class OpenSearchRestTest {
             .build();
   }
 
-  @AfterAll
-  static void tearDownOpenSearch() {
-    opensearch.stop();
-  }
-
-  @Test
-  void shouldGetStatusWithTraces() throws IOException {
-    client.performRequest(new Request("GET", "_cluster/health"));
-
-    testing.waitAndAssertTraces(
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span ->
-                    span.hasName("GET")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(maybeStable(DB_SYSTEM), "opensearch"),
-                            equalTo(maybeStable(DB_OPERATION), "GET"),
-                            equalTo(maybeStable(DB_STATEMENT), "GET _cluster/health")),
-                span ->
-                    span.hasName("GET")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasParent(trace.getSpan(0))
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(NETWORK_PROTOCOL_VERSION, "1.1"),
-                            equalTo(SERVER_ADDRESS, httpHost.getHostName()),
-                            equalTo(SERVER_PORT, httpHost.getPort()),
-                            equalTo(HTTP_REQUEST_METHOD, "GET"),
-                            equalTo(URL_FULL, httpHost.toURI() + "/_cluster/health"),
-                            equalTo(HTTP_RESPONSE_STATUS_CODE, 200L))));
-  }
-
-  @Test
-  void shouldGetStatusAsyncWithTraces() throws Exception {
-    AtomicReference<Response> requestResponse = new AtomicReference<>(null);
-    AtomicReference<Exception> exception = new AtomicReference<>(null);
-    CountDownLatch countDownLatch = new CountDownLatch(1);
-
-    ResponseListener responseListener =
-        new ResponseListener() {
-          @Override
-          public void onSuccess(Response response) {
-            testing.runWithSpan(
-                "callback",
-                () -> {
-                  requestResponse.set(response);
-                  countDownLatch.countDown();
-                });
-          }
-
-          @Override
-          public void onFailure(Exception e) {
-            testing.runWithSpan(
-                "callback",
-                () -> {
-                  exception.set(e);
-                  countDownLatch.countDown();
-                });
-          }
-        };
-
-    testing.runWithSpan(
-        "client",
-        () -> {
-          client.performRequestAsync(new Request("GET", "_cluster/health"), responseListener);
-        });
-    countDownLatch.await();
-
-    if (exception.get() != null) {
-      throw exception.get();
-    }
-
-    testing.waitAndAssertTraces(
-        trace ->
-            trace.hasSpansSatisfyingExactly(
-                span -> span.hasName("client").hasKind(SpanKind.INTERNAL),
-                span ->
-                    span.hasName("GET")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasParent(trace.getSpan(0))
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(maybeStable(DB_SYSTEM), "opensearch"),
-                            equalTo(maybeStable(DB_OPERATION), "GET"),
-                            equalTo(maybeStable(DB_STATEMENT), "GET _cluster/health")),
-                span ->
-                    span.hasName("GET")
-                        .hasKind(SpanKind.CLIENT)
-                        .hasParent(trace.getSpan(1))
-                        .hasAttributesSatisfyingExactly(
-                            equalTo(NETWORK_PROTOCOL_VERSION, "1.1"),
-                            equalTo(SERVER_ADDRESS, httpHost.getHostName()),
-                            equalTo(SERVER_PORT, httpHost.getPort()),
-                            equalTo(HTTP_REQUEST_METHOD, "GET"),
-                            equalTo(URL_FULL, httpHost.toURI() + "/_cluster/health"),
-                            equalTo(HTTP_RESPONSE_STATUS_CODE, 200L)),
-                span ->
-                    span.hasName("callback")
-                        .hasKind(SpanKind.INTERNAL)
-                        .hasParent(trace.getSpan(0))));
+  @Override
+  protected int getResponseStatus(Response response) {
+    return response.getStatusLine().getStatusCode();
   }
 }
