+++ instrumentation/opensearch/opensearch-rest-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/opensearch/rest/v3_0/OpenSearchRestSingletons.java
@@ -26,7 +26,6 @@ public final class OpenSearchRestSingletons {
 
       @Override
       public int getStatusCode() {
-        new Exception().printStackTrace();
         return response.getStatusLine().getStatusCode();
       }
